Move OptimizeProcess to WPF to remove dependency
